style: apply cargo fmt + clippy fixes across entire workspace
- cargo fmt --all: consistent formatting across all 139 source files
- cargo clippy --fix: auto-fix collapsible if, deref patterns, format args, etc.
- Manual fixes: clamp() instead of .max(0).min(100), #[allow] for legitimate too_many_arguments
- Fixed empty line after doc comment in theme.rs
- Renamed simulation_tests inner module to avoid module_inception warning
- Suppressed field_reassign_with_default in test functions
- Result: 0 clippy warnings, 0 fmt diffs, 215 tests passing

+++ crates/osl-app/src/app/simulation/history.rs
@@ -54,7 +54,7 @@ impl SimulationHistoryEntry {
             analysis_type: analysis_type.to_string(),
             analysis_body: analysis_body.to_string(),
             backend: backend.to_string(),
-            duration_ms: run.metadata.duration_ms as u128,
+            duration_ms: run.metadata.duration_ms,
             status_str: run.metadata.status.as_str().to_string(),
             output_dir: run.output_dir.display().to_string(),
             settings_summary,
@@ -178,7 +178,6 @@ impl SimulationHistory {
         Self { entries }
     }
 }
-
 
 /// Path to the history JSON file.
 fn history_path() -> PathBuf {
